reorganizing with modules
much nicer, as now my html file only needs to reference leaflet itself, as the index.js.  no longer needs reference to l.geometryutil or leaflet-vectorhats.js in the html.  however, this required a a mod to the l.geometryutil source, to allow an export, which is then imported into my leaflet-vectorhats.js.  opened an issue with geometryutil to see if that line can be added.