generalizing
trying to generalize the pushhat function to take in size, coordinates (as an array) and bearings (as an array), which will then build all the vectorhats.  the size and coords and bearings will be pre-calculated through the if else tree based on the user-input options.  good news is that a options.frequency = 1 is the same as 'endonly' so I can remove the index as an argument for the pushhats function